build process should watch all the relevant files and run the build in dependency order
