+++ workspace-scripts/notify.sh
@@ -1,0 +1,4 @@
+MESSAGE=$1
+osascript -e "display notification \"${MESSAGE}\""
+# say "$MESSAGE"
+
